# be-lspmpi/Dto/ChangePasswordRequest.cs
namespace be_lspmpi.Dto
{
    public class ChangePasswordRequest
    {
        public string OldPassword { get; set; } = string.Empty;
        public string NewPassword { get; set; } = string.Empty;
    }
}