# be-lspmpi/Dto/RegisterRequest.cs
namespace be_lspmpi.Dto
{
    public class RegisterRequest
    {
        public string? Username { get; set; }
        public string? Password { get; set; }
        public string? Name { get; set; }
        public string? Email { get; set; }
    }
}