# be-lspmpi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lemdiklat Permapendis API",
    "version": "1.0.0",
    "description": "Complete API for Lemdiklat Permapendis system"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user and create session cookie",
        "operationId": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User registration",
        "description": "Create new user account",
        "operationId": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User logout",
        "description": "Clear user session cookie",
        "operationId": "Logout",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Change user password",
        "description": "Update current user password",
        "operationId": "ChangePassword",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Forgot password",
        "description": "Send password reset instructions to email",
        "operationId": "ForgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "email"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset instructions sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "description": "Get authenticated user information from cookie",
        "operationId": "GetCurrentUser",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClaimUser" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/users/find": {
      "post": {
        "tags": ["Users"],
        "summary": "Find users",
        "description": "Search users with criteria",
        "operationId": "FindUsers",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FindRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Users found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "description": "Retrieve user information by user ID",
        "operationId": "GetUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "description": "Bad request" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user profile",
        "description": "Update user profile information",
        "operationId": "UpdateUser",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserProfile" }
            }
          }
        },
        "responses": {
          "204": { "description": "User updated successfully" },
          "400": { "description": "Bad request" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "description": "Delete user account",
        "operationId": "DeleteUser",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "User deleted successfully" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Create new user",
        "description": "Create a new user account (Admin and Manager only)",
        "operationId": "CreateUser",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "403": { "description": "Forbidden - insufficient permissions" }
        }
      }
    },
    "/users/profile-photo": {
      "post": {
        "tags": ["Users"],
        "summary": "Upload profile photo",
        "description": "Upload user profile photo",
        "operationId": "UploadProfilePhoto",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["photo"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile photo uploaded successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PhotoUploadResponse" }
              }
            }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/users/profile-photo/{userId}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get profile photo",
        "description": "Retrieve user profile photo",
        "operationId": "GetProfilePhoto",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile photo retrieved successfully",
            "content": {
              "image/jpeg": { "schema": { "type": "string", "format": "binary" } },
              "image/png": { "schema": { "type": "string", "format": "binary" } },
              "image/webp": { "schema": { "type": "string", "format": "binary" } },
              "image/svg+xml": { "schema": { "type": "string", "format": "binary" } }
            }
          },
          "400": { "description": "Bad request" },
          "404": { "description": "Profile photo not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "nullable": true },
          "password": { "type": "string", "nullable": true }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "isAuth": { "type": "boolean" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "nullable": true },
          "password": { "type": "string", "nullable": true },
          "name": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "oldPassword": { "type": "string" },
          "newPassword": { "type": "string" }
        },
        "required": ["oldPassword", "newPassword"]
      },
      "ServiceResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" }
        }
      },
      "ClaimUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "nullable": true },
          "roleId": { "type": "integer", "format": "int32" },
          "username": { "type": "string", "nullable": true },
          "profileName": { "type": "string", "nullable": true }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string", "maxLength": 50 },
          "password": { "type": "string", "maxLength": 255 },
          "passwordSalt": { "type": "string", "maxLength": 255 },
          "roleId": { "type": "integer", "format": "int32" },
          "role": { "$ref": "#/components/schemas/Role" },
          "userProfileId": { "type": "string", "format": "uuid" },
          "userProfile": { "$ref": "#/components/schemas/UserProfile" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "isActivated": { "type": "boolean" }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "name": { "type": "string", "maxLength": 50, "nullable": true },
          "level": { "type": "integer", "format": "int32" }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "maxLength": 100 },
          "email": { "type": "string", "maxLength": 255 },
          "phone": { "type": "string", "maxLength": 20, "nullable": true },
          "address": { "type": "string", "maxLength": 500, "nullable": true },
          "city": { "type": "string", "maxLength": 100, "nullable": true },
          "profilePicture": { "type": "string", "maxLength": 255, "nullable": true }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["name", "email", "username", "password"]
      },
      "PhotoUploadResponse": {
        "type": "object",
        "properties": {
          "fileName": { "type": "string" },
          "message": { "type": "string" }
        }
      },
...
      "FindRequest": {
        "type": "object",
        "properties": {
          "search": { "type": "string" },
          "page": { "type": "integer", "format": "int32", "default": 1 },
          "pageSize": { "type": "integer", "format": "int32", "default": 10 },
          "sort": { "type": "string" },
          "order": { "type": "string" },
          "filter": { "type": "string" }
        }
      },
...
    }
  }
}